Move classes into a separate directory
