+++ tools/profiler/profiler/preprocess.py
@@ -2,7 +2,7 @@ import json
 import os
 
 from collections import defaultdict
-from profiler.classes import CellMetadata, ControlMetadata, ParChildType
+from profiler.classes.classes import CellMetadata, ControlMetadata, ParChildType
 
 
 def read_shared_cells_map(shared_cells_json) -> dict[str, dict[str, str]]:
