+++ tools/profiler/profiler/visuals/utilization_plots.py
@@ -1,7 +1,7 @@
 from collections import Counter
 import pandas as pd
 import plotly.express as px
-from profiler.classes import UtilizationCycleTrace
+from profiler.classes.classes import UtilizationCycleTrace
 
 
 class Plotter:
